backend/pyproject.toml
[project]
name = "coderefine-backend"
version = "0.1.0"
description = "Agentic AI Backend"
authors = [
    {name = "Your Name"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "google-genai (>=1.56.0,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
